ADDED Status effect
Currently supports stuns, slows and damage over time, with fear and confusion planned.

+++ StortSpelprojekt/StortSpelprojekt/GameObjects/SecurityCamera.cpp
@@ -66,7 +66,7 @@ void SecurityCamera::CheckVisibleTiles()
 	Enemy* test = nullptr;
 	//_visionCone->ColorVisibleTiles({0,0,3});
 	AI::Vec2D tile = {0,0};
-	for (size_t i = 0; i < _visionCone->GetNrOfVisibleTiles(); i++)
+	for (int i = 0; i < _visionCone->GetNrOfVisibleTiles(); i++)
 	{
 		tile = _visionCone->GetVisibleTile(i);
 		if (_tileMap->IsEnemyOnTile(tile))
